AI or ML Engineering/lpdp-rag/requirements.txt
# Web Framework
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0

# LangChain Environment - Stable versions
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.39
langchain-groq==0.1.9
langgraph==0.2.28
langsmith==0.1.129

# Vector embeddings and database
langchain-chroma==0.1.4
langchain-huggingface==0.0.3
chromadb==0.5.11
sentence-transformers==3.1.1

# Document Processing
pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
markdown==3.7

# Web Scraping
requests==2.32.3

# Translation Service
deep-translator==1.11.4

# Utilities & Core Dependencies
numpy==1.26.4
pydantic==2.9.2
typing-extensions==4.12.2

# Text Splitters
langchain-text-splitters==0.2.4

# Additional dependencies
PyPDF2==3.0.1